+++ main/src/main/java/org/teachmeskills/MainApplication.java
@@ -9,3 +9,4 @@ public class MainApplication {
     SpringApplication.run(MainApplication.class, args);
   }
 }
+
