+++ main/src/test/java/tenderTests/OrganizationServiceTest.java
@@ -1,3 +1,5 @@
+package tenderTests;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -69,25 +71,6 @@ public class OrganizationServiceTest {
     assertEquals(expected, created);
   }
 
-//  @Test
-//  public void createOrganizationTest() {
-//
-//    CreateOrganizationDto createOrganizationDto = new CreateOrganizationDto();
-//    createOrganizationDto.setUnp(222222222);
-//    createOrganizationDto.setFullName("Общество с ограниченной ответственностью 'Ромашка'");
-//    createOrganizationDto.setShortName("ООО 'Ромашка'");
-//    createOrganizationDto.setLegalAddress("г. Минск");
-//    createOrganizationDto.setActualAddress("г. Минск");
-//
-//    Organization created = new Organization(2, 222222222,
-//        "Общество с ограниченной ответственностью 'Ромашка'",  "ООО 'Ромашка'",
-//        "г. Минск","г. Минск",null,null,null,null);
-//
-//    given(organizationService.createOrganization(createOrganizationDto)).willReturn(created);
-//    Organization expected = organizationService.createOrganization(createOrganizationDto);
-//
-//    assertEquals(expected, created);
-//  }
 
   @Test
   public void findOrganizationsByUsersIsNullTest() {
@@ -110,24 +93,4 @@ public class OrganizationServiceTest {
     assertEquals(expected, organizationList);
   }
 
-  @Test
-  public void findPaginatedAllOrganizationsTest() {
-
-    List<Organization> organizationList = new ArrayList<>();
-    organizationList.add(new Organization (2, 222222222,"Общество с ограниченной ответственностью 'Ромашка'","ООО 'Ромашка'", " г. Минск",
-        "г. Минск",null,null,null,null));
-    organizationList.add(new Organization (3, 333333333,"Общество с ограниченной ответственностью 'Роза'", "ООО 'Роза'", "г. Минск",
-        "г. Минск", null,null,null,null));
-    organizationList.add(new Organization (4,444444444 ,"Общество с ограниченной ответственностью 'Пион'", "ООО 'Пион'", " г. Минск",
-        "г. Минск",null,null,null,null));
-    organizationList.add(new Organization (5,555555555, "Общество с ограниченной ответственностью 'МТЗ'", "ООО 'МТЗ'", " г. Минск",
-        "г. Минск",null,null,null,null));
-    organizationList.add(new Organization (6,666666666, "Общество с ограниченной ответственностью 'Geely'","ООО 'Geely'", " г. Минск",
-        "г. Минск", null,null,null,null ));
-
-    Page<Organization> paginated = organizationRepository.findAll(PageRequest.of(2, 3));
-    Page<Organization> paginatedUsers = organizationService.findPaginatedAllOrganizations(2, 3);
-
-    assertEquals(paginatedUsers, paginated);
-  }
 }
